Settings: Forward port TapToWake
original commit can be found here
https://github.com/CyanogenMod/android_packages_apps_Settings/commit/ded2213ceebd1ff5bece7b9e9846551d804f4d51

Contains Commits:

Author: Ricardo Cerqueira<cyanogenmod@cerqueira.org>
DisplaySettings: Add hardware-framework support for tap-to-wake

Since we have tap-to-sleep globally, might as well support the
wake gesture as a standalone thing through the HAF. Place it in
"Display".

Change-Id: I6facc2334ff9e80077581c54c428476594e91528

Author: Roman Birg <roman@cyngn.com>
Settings: set proper default value for tap to wake

The first time the display_settings.xml file gets inflated, the
PreferenceManager takes the defaultValue and persists it in the default
shared preferences file.
When we read the double tap to wake preference,
we assume it has a default value of true, which it uses until the user
enters Display Settings for the first time. At that point, the default
value is persisted, and every boot after will use the default value in
display_settings.xml.
Remove the default defined in the XML file.

Change-Id: I0989dd73b6b3c42ff8649bc17eebf4e191293ee4

Change-Id: I271390aa7bfc56811dcf6fff7e1d05c76c4bfa48